LayoutTests/editing/deleting/5026848-2-expected.txt

This tests to see that selected non-editable nodes that are not in both the start and end editable root are preserved during a deletion.  You should see two 'Non-editable' pieces on either side of the editable piece below.
| "
"
| <table>
|   contenteditable="false"
|   "
    "
|   <tbody>
|     <tr>
|       "
        "
|       <td>
|         "Non-editable"
|       "
        "
|       <td>
|         contenteditable="true"
|         "Editable <#selection-caret>"
|         "content"
|       "
        "
|       <td>
|         "Non-editable"
|       "
    "
|     "
"
